+++ packages/eslint-plugin-template/src/rules/conditional-complexity.ts
@@ -124,12 +124,10 @@ function getTotalComplexity(ast: AST): number {
   const possibleBinaryOrConditional =
     extractPossibleBinaryOrConditionalFrom(ast);
 
-  if (
-    !(
-      possibleBinaryOrConditional instanceof Binary ||
-      possibleBinaryOrConditional instanceof Conditional
-    )
-  ) {
+  if (!(
+    possibleBinaryOrConditional instanceof Binary ||
+    possibleBinaryOrConditional instanceof Conditional
+  )) {
     return 0;
   }